requirements.txt
numpy==1.23.4
numba==0.56.4
scipy==1.9.3
scikit-learn==1.1.3
faiss-gpu==1.7.2
umap-learn==0.5.3
matplotlib==3.5.3
pandas==1.4.4